added double check for data deletion

+++ js/controls.js
@@ -2,9 +2,10 @@
 
 
 // test array
-// const profileArray = [{kidName : 'John', color : 'red', animal : 'Tiger', number: 5, timesVideoWatched : 0, timesPlayedPhysics: 0},{kidName : 'Alice', color : 'blue', animal : 'bear', number: 3, timesVideoWatched : 2, timesPlayedPhysics: 6},{kidName : 'Bob', color : 'green', animal : 'bird', number: 3, timesVideoWatched : 1, timesPlayedPhysics: 36}];
+const profileArray = [{kidName : 'John', color : 'red', animal : 'Tiger', number: 5, timesVideoWatched : 0, timesPlayedPhysics: 0},{kidName : 'Alice', color : 'blue', animal : 'bear', number: 3, timesVideoWatched : 2, timesPlayedPhysics: 6},{kidName : 'Bob', color : 'green', animal : 'bird', number: 3, timesVideoWatched : 1, timesPlayedPhysics: 36}];
 
-const profileArray = JSON.parse(localStorage.getItem(profileArray));
+// pull in localStorage
+// const profileArray = JSON.parse(localStorage.getItem(profileArray));
 const progressChart = document.getElementById('myChart');
 
 console.log(profileArray);
@@ -50,7 +51,13 @@ const resetDataEl = document.getElementById('delete-data');
 resetDataEl.addEventListener('click', clearData);
 
 function clearData(event){
-  localStorage.clear();
-  alert('Data has been erased!');
-  window.location.replace("splash.html");
+  const response = confirm('Are you sure you want to clear all data?');
+  if (response){
+    localStorage.clear();
+    alert('Data has been erased!');
+    window.location.replace('splash.html');
+  }
+  else {
+    alert('Data was not deleted.');
+  }
 }
